requirements.txt
aiogram-calendar==0.5.0
python-dateutil==2.9.0.post0
aiogram~=3.6.0
SQLAlchemy~=2.0.30
APScheduler~=3.10.4
python-dotenv==1.0.1
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
alembic==1.13.1
psycopg2-binary==2.9.9